+++ mechanicskit/__init__.py
@@ -49,7 +49,7 @@ from .animation_utils import to_responsive_html
 from .markdown import md, Markdown
 
 # Version information
-__version__ = '0.7.0'
+__version__ = '0.7.1'
 
 def version():
     """